+++ CHANGES.md
@@ -1,13 +1,13 @@
 Full changelog
 ==============
 
-v1.4.0 (unreleased)
+v1.4.0 (2022-05-31)
 -------------------
 
 * Add support for specifying visual attributes when creating a subset group. [#2297]
 
 * Modify profile viewer so that when in 'Sum' mode, parts of profiles
-  with no valid values are NaN rather than zero.
+  with no valid values are NaN rather than zero. [#2298]
 
 * Add support for using degrees in full-sphere projections. [#2279]